1. 差买票的充分调试 2. 注册   2.1 身份证校验 3. 登陆拦截   3.1 登陆验证码

+++ src/main/java/com/air/service/UserService.java
@@ -10,7 +10,7 @@ import java.util.List;
  */
 public interface UserService {
 
-    int saveUser(User user);
+    Long saveUser(User user);
 
     void updateUser(User user);
